% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_defined_activities.R
\name{get_defined_activities}
\alias{get_defined_activities}
\title{Get valid activities}
\usage{
get_defined_activities()
}
\value{
a character vector of recognizable activity classes
}
\description{
Get valid activities
}
