% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_actor_types.R
\name{get_project_actor_type_actions}
\alias{get_project_actor_type_actions}
\title{get_project_actor_type_actions}
\usage{
get_project_actor_type_actions(project_file, actor_types = NULL)
}
\arguments{
\item{project_file}{character vector of length one with the path to the
project file.}

\item{actor_types}{character vector of one or more actor types. If NULL,
all actor types are used.}
}
\value{
A list of actor_types where each has a character vector with their
respective actions.
}
\description{
get_project_actor_type_actions
}
\examples{
project_file <- system.file("extdata", "project_yaml_files",
                            "example_colony_project.yml",
                            package = "crosssectiontimer", lib.loc = NULL,
                            mustWork = FALSE)
actor_types <- get_project_actor_types(project_file)
}
