% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_actor_types.R
\name{get_project_actor_types}
\alias{get_project_actor_types}
\title{get_project_actor_types}
\usage{
get_project_actor_types(project_file)
}
\arguments{
\item{project_file}{character vector of length one with the path to the
project file.}
}
\value{
A character vector with the actor types.
}
\description{
get_project_actor_types
}
\examples{
project_file <- system.file("extdata", "project_yaml_files",
                            "example_colony_project.yml",
                            package = "crosssectiontimer", lib.loc = NULL,
                            mustWork = FALSE)
actor_types <- get_project_actor_types(project_file)
}
