% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_activity_observation.R
\name{make_activity_observation}
\alias{make_activity_observation}
\title{Make a sampled activity observation}
\usage{
make_activity_observation(activities, size = 1, prob)
}
\arguments{
\item{activities}{data.frame object with a single actor_type}

\item{size}{the number of actors * n_months * time_per_month}

\item{prob}{a vector of probability weights for obtaining the elements
of the vector being sampled.}
}
\value{
an activity observation object
}
\description{
Make a sampled activity observation
}
