% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_complete_sample_durations.R
\name{make_complete_sample_durations}
\alias{make_complete_sample_durations}
\title{Make a complete sample duration data.frame object}
\usage{
make_complete_sample_durations(sample_duration, actor_activities)
}
\arguments{
\item{sample_duration}{named numeric vector that may be missing actions if
those actions did not occur during the sampling phase.}

\item{actor_activities}{data.frame of all possible actor actions.}
}
\value{
an sample duration data.frame object
}
\description{
Adds back missing activities that were dropped by the \code{table} function
because they do not appear in the sample. They are added into the object
and given a count of zero.
}
