% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sample_duration.R
\name{sim_sample_duration}
\alias{sim_sample_duration}
\title{Simulation of actors activities to estimate duration on each activity}
\usage{
sim_sample_duration(
  activities,
  size = n_actors * times_per_month * n_months,
  iterations = 1
)
}
\arguments{
\item{activities}{data.frame object with a single actor_type}

\item{size}{the number of actors * n_months * time_per_month}

\item{iterations}{integer indicating the number of times the simulation
is to occur.}
}
\description{
The simulation repeats the estimation experiment to allow estimation of the
precision of the estimates expected from the estimation experiment.
}
